Alteração no banco de dados para estabelecer conexão, implementação de cadastro e listagem de dados.

+++ src/ProdutosDTO.java
@@ -10,7 +10,7 @@
 public class ProdutosDTO {
     private Integer id;
     private String nome;
-    private Integer valor;
+    private Double valor; // Alterei para Double
     private String status;
 
     public Integer getId() {
@@ -21,7 +21,7 @@ public class ProdutosDTO {
         return nome;
     }
 
-    public Integer getValor() {
+    public Double getValor() { // Alterei para Double
         return valor;
     }
 
@@ -37,12 +37,11 @@ public class ProdutosDTO {
         this.nome = nome;
     }
 
-    public void setValor(Integer valor) {
+    public void setValor(Double valor) { // Alterei para Double
         this.valor = valor;
     }
 
     public void setStatus(String status) {
         this.status = status;
     }
-    
 }
